+++ parser/optionsBase.h
@@ -36,9 +36,10 @@ namespace GMAD
   public:
     OptionsBase();
 
-    std::string inputFileName;    ///< input filename
-    std::string visMacroFileName; ///< visualisation filename
-    bool        visDebug;         ///< flag for visualisation debug
+    std::string inputFileName;       ///< Input filename.
+    std::string visMacroFileName;    ///< Visualisation filename.
+    std::string geant4MacroFileName; ///< Geant4 macro to run.
+    bool        visDebug;            ///< Flag for visualisation debug.
   
     ///@{ Parameter for output format
     std::string outputFileName;