+++ manual/source/version_history.rst
@@ -26,9 +26,10 @@ New Features
   particle used for magnetic field strength calculations (:code:`beamParticleName`).
 * Specify the particle assumed for the user file distribution that can be different from
   the design particle.
-  
 * New option to use a one turn map generated from MAD-X PTC to correct
   multi-turn tracking for circular machines.
+* New option :code:`geant4Macro` and executable option :code:`--geant4Macro` to run an optional
+  macro in the visualiser once it's started.
 
 General
 -------
@@ -47,6 +48,8 @@ General
   as this requires G4MultiSensitiveDetector.
 * New tests for testing backwards compatibility of analysis tool with previous data version.
 * "Model Preparation" is now "Model Conversion" in the manual to be clearer.
+* Visualisation now uses macro search path to look for visualisation macro in the installation
+  directory then the build directory of BDSIM.
   
 Bug Fixes
 ---------